pyproject.toml
[project]
name = "flex-carbon-copy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "gradio>=5.29.1",
    "pydantic>=2.11.4",
    "pystray>=0.19.5",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0.2",
    "win11toast>=0.35",
]
